Rename parts data entrypoints
