+++ docs/features/chat.md
@@ -22,8 +22,8 @@ Enables guided self-reflection, parts work, and agent-assisted workflows.
 ## How it works
 - UI at app/chat/page.tsx with streaming responses
 - useChat hook manages message state and streaming
-- Client data access uses `@/lib/data/parts-lite` (browser-safe)
-- Server routes/actions use `@/lib/data/parts-server` for writes, logging, and snapshots
+- Client data access uses `@/lib/data/parts-client` (browser-safe)
+- Server routes/actions use `@/lib/data/parts.server` for writes, logging, and snapshots
 - Agent actions are logged via lib/database/action-logger.ts (server-only)
 
 ## Data model
